requirements.txt
alembic==1.11.2
antiorm==1.2.1
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
db==0.1.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.3
Flask==2.3.2
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fsspec==2023.9.0
greenlet==2.0.2
gunicorn==21.2.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.25.1
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.5.0
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
setuptools==68.0.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.15
stack-data==0.6.2
threadpoolctl==3.2.0
tokenizers==0.13.3
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.4
wheel==0.38.4
WTForms==3.0.1
xmltodict==0.13.0
zipp==3.15.0
